Add dependency facet
Any dependency can have a facet now. By default, a dependency flagged
as a 'web' facet brings the necessary infrastructure to start a  web
container. If no such facet is defined for a war project, the web starter
is added automatically as before except we don't hard code the list of
candidates in the code anymore