Expose full configuration meta-data
InitializrMetadataBuilder can now merge a number of resources into a
single meta-data instance. A new `/config` endpoint is now available and
exposes the configuration meta-data of the current instance.

Combining those two, it is now possible to bootstrap an instance without
a single line of configuration; instead, the meta-data are built from the
content of a remote service.

Each capability has now a `merge` method with a "last wins" strategy. For
collections, unknown elements are added.

Closes gh-88