Polish contribution
`1.3.2.RELEASE` is available now so there is no need to have a version
range on the snapshot.  We don't need to copy/paste the version of the
bom in the actual dependency (that's the whole business of the bom
definition (including versions mapping according to the Spring Boot
version).

We want the `cloud-bom` to be automatically added when the data flow bom
is required so it is referenced as "additional BOM" as well.

Closes gh-189