Support of additional BOM
BOMs are structured in such a way than adding one to the project may
require another one(s) to be added to the project.

This commit adds an `additionalBoms` property to `BillOfMaterials` that
can be used to refer the BOM(s) that should be automatically added if
said BOM is added to the project.

Closes gh-190