Polish contribution
This commit makes sure that each dependency link is HAL compliant (like
the project types in the metadata). Links are grouped by relations with
well known relations to be defined (i.e. 'how-to', 'reference', 'home'
and so forth).

Each link can be "templated" (in the HAL sense) and only `{bootVersion}`
is supported at the moment. This is useful is a precise anchor should
reference to the actual Stpring Boot version chosen by the user. If the
url contains `{bootVersion}` it is automatically flagged as being
templated.

The UI endpoint provides those links at the moment but this may become
unecessary as it isn't sure yet how those links are going to be managed
by the web client.

Closes gh-279